chore: move pd tests to hg-pd-test
