+++ hugegraph-pd/pom.xml
@@ -143,176 +143,6 @@
                     </filesets>
                 </configuration>
             </plugin>
-            <!-- Apache RAT for license check -->
-            <plugin>
-                <groupId>org.apache.rat</groupId>
-                <artifactId>apache-rat-plugin</artifactId>
-                <configuration>
-                    <excludes>
-                        <exclude>**/*.versionsBackup</exclude>
-                        <exclude>**/*.proto</exclude>
-                        <exclude>**/*.log</exclude>
-                        <exclude>**/*.txt</exclude>
-                        <exclude>**/*.json</exclude>
-                        <exclude>**/*.conf</exclude>
-                        <exclude>**/*.map</exclude>
-                        <exclude>**/*.properties</exclude>
-                        <exclude>dist/**/*</exclude>
-                        <exclude>**/assembly/static/bin/hugegraph.service</exclude>
-                        <exclude>**/swagger-ui/**/*</exclude>
-                        <exclude>scripts/dev/reviewers</exclude>
-                        <exclude>scripts/dev/reviewers</exclude>
-                        <exclude>**/*.md</exclude>
-                        <exclude>**/dependency-reduced-pom.xml</exclude>
-                        <exclude>**/logs/*.log</exclude>
-                        <exclude>**/META-INF/**/*</exclude>
-                        <exclude>**/target/*</exclude>
-                        <exclude>style/*</exclude>
-                        <exclude>ChangeLog</exclude>
-                        <exclude>CONFIG.ini</exclude>
-                        <exclude>GROUPS</exclude>
-                        <exclude>OWNERS</exclude>
-                        <exclude>**/grpc/**</exclude>
-                        <!-- Git & GitHub  -->
-                        <exclude>.github/**/*</exclude>
-                        <exclude>.gitignore</exclude>
-                        <exclude>.gitattributes</exclude>
-                        <!-- Intellij -->
-                        <exclude>**/*.iml</exclude>
-                        <exclude>**/*.iws</exclude>
-                        <exclude>**/*.ipr</exclude>
-                        <exclude>**/META-INF/MANIFEST.MF</exclude>
-                        <!-- Maven -->
-                        <exclude>.repository/**</exclude>
-                        <exclude>**/.flattened-pom.xml</exclude>
-                    </excludes>
-                    <consoleOutput>true</consoleOutput>
-                </configuration>
-            </plugin>
         </plugins>
     </build>
-
-    <profiles>
-        <profile>
-            <id>client-test</id>
-            <activation>
-                <activeByDefault>true</activeByDefault>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>2.20</version>
-                        <executions>
-                            <execution>
-                                <id>client-test</id>
-                                <goals>
-                                    <goal>test</goal>
-                                </goals>
-                                <phase>test</phase>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <id>core-test</id>
-            <activation>
-                <activeByDefault>true</activeByDefault>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>2.20</version>
-                        <executions>
-                            <execution>
-                                <id>core-test</id>
-                                <goals>
-                                    <goal>test</goal>
-                                </goals>
-                                <phase>test</phase>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <id>cli-tools-test</id>
-            <activation>
-                <activeByDefault>true</activeByDefault>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>2.20</version>
-                        <executions>
-                            <execution>
-                                <id>cli-tools-test</id>
-                                <goals>
-                                    <goal>test</goal>
-                                </goals>
-                                <phase>test</phase>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <id>common-test</id>
-            <activation>
-                <activeByDefault>true</activeByDefault>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>2.20</version>
-                        <executions>
-                            <execution>
-                                <id>common-test</id>
-                                <goals>
-                                    <goal>test</goal>
-                                </goals>
-                                <phase>test</phase>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <id>service-test</id>
-            <activation>
-                <activeByDefault>true</activeByDefault>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>2.20</version>
-                        <executions>
-                            <execution>
-                                <id>service-test</id>
-                                <goals>
-                                    <goal>test</goal>
-                                </goals>
-                                <phase>test</phase>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
-
 </project>
