+++ hugegraph-server/hugegraph-dist/src/assembly/travis/run-api-test-for-raft.sh
@@ -57,7 +57,7 @@ export HUGEGRAPH_PASSWORD=pa
 $RAFT_TOOLS --set-leader "hugegraph" "$RAFT_LEADER"
 
 # run api-test
-mvn test -pl hugegraph-server -am -P api-test,$BACKEND || (cat $RAFT1_DIR/logs/hugegraph-server.log && exit 1)
+mvn test -pl hugegraph-server/hugegraph-test -am -P api-test,$BACKEND || (cat $RAFT1_DIR/logs/hugegraph-server.log && exit 1)
 
 $TRAVIS_DIR/build-report.sh $BACKEND $JACOCO_PORT $REPORT_FILE
