+++ hugegraph-server/hugegraph-dist/src/assembly/travis/run-tinkerpop-test.sh
@@ -21,9 +21,9 @@ BACKEND=$1
 SUITE=$2
 
 if [[ "$SUITE" == "structure" || "$SUITE" == "tinkerpop" ]]; then
-    mvn test -pl hugegraph-server -am -P tinkerpop-structure-test,$BACKEND
+    mvn test -pl hugegraph-server/hugegraph-test -am -P tinkerpop-structure-test,$BACKEND
 fi
 
 if [[ "$SUITE" == "process" || "$SUITE" == "tinkerpop" ]]; then
-    mvn test -pl hugegraph-server -am -P tinkerpop-process-test,$BACKEND
+    mvn test -pl hugegraph-server/hugegraph-test -am -P tinkerpop-process-test,$BACKEND
 fi
